arregle errores

+++ src/main/java/com/aquienllamo/aquienllamo/model/auth/repositories/CredentialsRepository.java
@@ -1,5 +1,6 @@
 package com.aquienllamo.aquienllamo.model.auth.repositories;
 
+
 import com.aquienllamo.aquienllamo.model.auth.Credentials.CredentialsEntity;
 import com.aquienllamo.aquienllamo.model.entities.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CredentialsRepository extends JpaRepository<CredentialsEntity, Long> {
+public interface CredentialsRepository extends JpaRepository<CredentialsEntity, Integer> {
 
     Optional<CredentialsEntity> findByUsername(String username);
     Optional<CredentialsEntity> findByRefreshToken(String refreshToken);
